+++ cmem_plugin_reason/plugin_validate.py
@@ -253,7 +253,7 @@ class ValidatePlugin(WorkflowPlugin):
             setup_cmempy_user_access(context.user)
             send_result(self.output_graph_iri, Path(self.temp) / "output.ttl")
             setup_cmempy_user_access(context.user)
-            post_provenance(self, get_provenance(self, context))
+            post_provenance(self, get_provenance(self, "Validate", context))
 
         valid_profiles = (
             self.add_profiles(validate_profiles(self, graphs)) if self.validate_profile else []